Add a HPTT path in transpose_op.cc
Summary:
https://arxiv.org/abs/1704.04374 is a simple, stateless library that
implements a high performance tensor transposition abstraction - it's
substantially faster than what we have. I think instead of going through an
engine specialization on the CPU side, we can just add this path, since there's
no value (in terms of state management, etc) for having it separate?

We could cache the plan, but it's so cheap to create in these tests.

Reviewed By: jonmorton

Differential Revision: D5534519

fbshipit-source-id: de2fd64fee11be259656b0f02f42a62b7035e3d3