Update observer sample rate
Summary:
We'd like to sparsely sample the net execution, but after the net is sampled for the first time, we'd like to densely sample the following few iterations so that we can have some meanful data for a short period of time.

Change the observer sample rate to the following:
skipIter: skip the first few iterations.
netInitSampleRate: the sample rate for the first iteration after the skipIter or immediately after reset.
netFollowupSampleRate: the sample rate after the netInitSampleRate is hit.
netFollowupSampleRate: the number of iterations that use the netFollowupSampleRate. After this number is hit, use netInitSampleRate (reset)
operatorNetSampleRatio: whenever the net is sampled, if the random number also hit operatorNetSampleRatio, collect operator metrics instead.

Reviewed By: Maratyszcza

Differential Revision: D6205657

fbshipit-source-id: da0c048f77fc4dc64f3fb71b6072429a57e9d2f0