Fix stats resporter in calulating STDDEV
Summary:
The STDDEV calculation code assumes that the `compare_exchange` returns the value of the atomic, while the C++ spec actually returns `bool`.

Also, the diff puts enough guard to avoid math error on python side -- although this should not happen, the guard is just to avoid problems with floating point calculation offsets.

Differential Revision: D6307930

fbshipit-source-id: d1754afb631f937aca7a88a82b5be2dd0c704aec